brings back OutputList spec
